Command Package View changelog entry #90 for details.
