+++ src/main/java/jterm/command/Exec.java
@@ -14,11 +14,13 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package main.java.jterm;
+package main.java.jterm.command;
 
 import java.io.*;
 import java.util.ArrayList;
 import java.util.Scanner;
+
+import main.java.jterm.JTerm;
 
 public class Exec
 {
