+++ src/main/java/jterm/command/Files.java
@@ -14,7 +14,7 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package main.java.jterm;
+package main.java.jterm.command;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -22,6 +22,8 @@ import java.io.FileWriter;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
+
+import main.java.jterm.JTerm;
 
 public class Files
 {
