+++ src/main/java/jterm/command/Ping.java
@@ -1,4 +1,4 @@
-package main.java.jterm;
+package main.java.jterm.command;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
